refactor(test): package 나누어서 test 코드 작성
