+++ src/test/java/baseball/domain/GameTest.java
@@ -1,0 +1,4 @@
+package baseball.domain;
+
+public class GameTest {
+}
