+++ src/test/java/baseball/view/ViewTest.java
@@ -1,0 +1,4 @@
+package baseball.view;
+
+public class ViewTest {
+}
